+++ src/components/App/Providers.tsx
@@ -4,8 +4,8 @@ import { ApolloProvider } from 'react-apollo';
 import { Provider } from 'react-redux';
 import { HashRouter } from 'react-router-dom';
 
-import store from './store';
-import client from './graphql/client';
+import store from '../../store';
+import client from '../../graphql/client';
 
 export const Providers = ({ children }: { children: React.ReactNode }) => (
   <HashRouter>
