+++ src/components/App/styles.css
@@ -1,4 +1,4 @@
-@import './colors.css';
+@import '../../colors.css';
 
 .grid {
   min-height: 100%;
